Working with Detail Screen

+++ data/src/main/java/com/sdk/data/remote/dto/Wind.kt
@@ -2,5 +2,5 @@ package com.sdk.data.remote.dto
 
 data class Wind(
     val deg: Int,
-    val speed: Int
+    val speed: Double
 )
